+++ elixir_tests/test/basic_funcionality_test.exs
@@ -3,37 +3,37 @@ defmodule BasicFuncionalityTest do
   #doctest BackwaterElixirTests
 
   test "expose and call erlang module" do
-    ref = start_backwater([{:erlang, [{:exports, :all}]}])
+    {ref, endpoint} = start_backwater([{:erlang, [{:exports, :all}]}])
 
     # call existing function
     assert {:ok, [1,2,3]} ==
-          :backwater_client.call(ref, :erlang, :binary_to_list, [<<1,2,3>>])
+      :backwater_client.call(endpoint, :erlang, :binary_to_list, [<<1,2,3>>])
 
     # call non-existing function
     assert {:error, {:remote, {:not_found, _headers, _body}}} =
-          :backwater_client.call(ref, :erlang, :binary_to_listsy, [<<1,2,3>>])
+      :backwater_client.call(endpoint, :erlang, :binary_to_listsy, [<<1,2,3>>])
 
     # call non-existing module
     assert {:error, {:remote, {:forbidden, _headers, _body}}} =
-          :backwater_client.call(ref, :erlangsy, :binary_to_list, [<<1,2,3>>])
+      :backwater_client.call(endpoint, :erlangsy, :binary_to_list, [<<1,2,3>>])
 
     stop_backwater(ref)
   end
 
   test "expose and call elixir module" do
-    ref = start_backwater([{Enum, [{:exports, :all}]}])
+    {ref, endpoint} = start_backwater([{Enum, [{:exports, :all}]}])
 
     # call existing function
     assert {:ok, [2,4,6]} ==
-          :backwater_client.call(ref, Enum, :map, [[1,2,3], fn (x) -> x * 2 end])
+      :backwater_client.call(endpoint, Enum, :map, [[1,2,3], fn (x) -> x * 2 end])
 
     # call non-existing function
     assert {:error, {:remote, {:not_found, _headers, _body}}} =
-          :backwater_client.call(ref, Enum, :mapsyy, [[1,2,3], fn (x) -> x * 2 end])
+      :backwater_client.call(endpoint, Enum, :mapsyy, [[1,2,3], fn (x) -> x * 2 end])
 
     # call non-existing module
     assert {:error, {:remote, {:forbidden, _headers, _body}}} =
-          :backwater_client.call(ref, Enumsyy, :map, [[1,2,3], fn (x) -> x * 2 end])
+      :backwater_client.call(endpoint, Enumsyy, :map, [[1,2,3], fn (x) -> x * 2 end])
 
     stop_backwater(ref)
   end
@@ -49,17 +49,13 @@ defmodule BasicFuncionalityTest do
                                       :decode_unsafe_terms => true },
                                     [{:port, 8080}],
                                     [])
-    # start client
-    :ok = :backwater_client.start(ref,
-                                  %{ :endpoint => "http://127.0.0.1:8080/",
-                                     :secret => secret,
-                                     :decode_unsafe_terms => true })
-    ref
+    endpoint = {"http://127.0.0.1:8080", secret}
+    {ref, endpoint}
   end
 
   def stop_backwater(ref) do
-    :ok = :backwater_client.stop(ref)
     :ok = :backwater_server.stop_listener(ref)
+    :timer.sleep(100)
   end
 
 end